+++ cmd/server/go.mod
@@ -1,4 +1,3 @@
 module server
 
 go 1.22.5
-
